scripts/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.26
azure-core==1.12.0
azure-identity==1.5.0
azure-mgmt-compute==19.0.0
azure-mgmt-core==1.2.2
azure-mgmt-network==17.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==16.0.0
azure-nspkg==3.0.2
backports.functools-lru-cache==1.6.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.3.2
cycler==0.10.0
enum34==1.1.10
idna==2.10
ipaddress==1.0.23
isodate==0.6.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
matplotlib==2.2.5
msal==1.10.0
msal-extensions==0.3.0
msrest==0.6.21
numpy==1.16.6
oauthlib==3.1.0
pandas==0.24.2
pathlib2==2.3.5
portalocker==1.7.1
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyxdg==0.25
requests==2.25.1
requests-oauthlib==1.3.0
scandir==1.10.0
scipy==1.2.3
seaborn==0.9.1
SecretStorage==2.3.1
subprocess32==3.5.4
typing==3.7.4.3
urllib3==1.26.3
